+++ src/pages/BasicAssessment.tsx
@@ -1,0 +1,7 @@
+import "../styles/basic-assessment.css";
+
+function BasicAssessment() {
+  return <h1>Basic Assessment Page</h1>;
+}
+
+export default BasicAssessment;
