+++ src/styles/basic-assessment.css
@@ -1,0 +1,3 @@
+body{
+    background-color: purple;
+}
